Fix ember-data canary dependency
Previously the ember-data dependency was *only* added to the
`package.json` file when *no node_module cache existed*. This caused
ember-data canary to be installed and correctly required for the *first*
fixture in tests. Subsequent tests would use whatever ember-data version
was specified by `ember new` as the requirement, which would then cause
a conflict, since the installed version was canary. This is fixed by
modifying the `package.json` for every pristine app generated, instead
of only the first one